Doctors verification requests

+++ Seravian/DTOs/Doctor/SendVerificationRequestResponseDto.cs
@@ -2,5 +2,5 @@ namespace Seravian.DTOs.Doctor;
 
 public class SendVerificationRequestResponseDto
 {
-    public int VerificationRequestId { get; set; }
+    public long VerificationRequestId { get; set; }
 }
